+++ sdk_core/command_handler/command_impl.cpp
@@ -759,6 +759,28 @@ livox_status CommandImpl::SetLivoxLidarRmcSyncTime(uint32_t handle, const char* 
                     MakeCommandCallback<LivoxLidarRmcSyncTimeResponse>(cb, client_data));
 }
 
+livox_status CommandImpl::SetLivoxLidarWorkModeAfterBoot(uint32_t handle, LivoxLidarWorkModeAfterBoot work_mode, LivoxLidarAsyncControlCallback cb, void* client_data) {
+  uint8_t req_buff[kMaxCommandBufferSize] = {0};
+  uint16_t req_len = 0;
+  
+  uint16_t key_num = 1;
+  memcpy(&req_buff[req_len], &key_num, sizeof(key_num));
+  req_len = sizeof(key_num) + sizeof(uint16_t);
+
+  LivoxLidarKeyValueParam * kv = (LivoxLidarKeyValueParam *)&req_buff[req_len];
+  kv->key = static_cast<uint16_t>(kKeyWorkModeAfterBoot);
+  kv->length = sizeof(uint8_t);
+  uint8_t* val_work_mode = reinterpret_cast<uint8_t*>(&kv->value[0]);
+  *val_work_mode = work_mode;
+  req_len += sizeof(LivoxLidarKeyValueParam) - 1 + sizeof(uint8_t);
+
+  return GeneralCommandHandler::GetInstance().SendCommand(handle,
+                    kCommandIDLidarWorkModeControl,
+                    req_buff,
+                    req_len,
+                    MakeCommandCallback<LivoxLidarAsyncControlResponse>(cb, client_data));
+}
+
 // Upgrade
 livox_status CommandImpl::LivoxLidarStartUpgrade(uint32_t handle, uint8_t *data, uint16_t length,
     LivoxLidarStartUpgradeCallback cb, void* client_data) {